refactor(transform/client): unify state-var assignment AST helpers
Combines three previously-separate AST helpers
(`state_simple_assigns_ast`, `state_compound_assigns_ast`,
`state_update_assigns_ast`) into one
`state_assigns_combined_ast` with a single visitor + a single
fixed-point loop sharing one parse + `SemanticBuilder` per
iteration.

Before: each of the three helpers did its own parse +
SemanticBuilder + visit + fixed-point (up to 16 iterations
apiece). State-var-heavy scripts could trigger up to ~48 parse
cycles just for these three concerns.

After: one parse + SemanticBuilder per fixed-point iteration.
Combined visitor handles AssignmentExpression (all operators)
and UpdateExpression in a single walk.

Each of the four call sites in mod.rs (module-script, reactive
statement, export let, instance script) collapses from 3
sequential helper calls to 1. The three obsolete helper files
are deleted (~1340 LOC).

Mapping preserved exactly — 12 unit tests in the new module
cover simple/compound/update individually + nested case +
runes proxy flag + shadow detection + parse-error / empty
inputs. Plus a new legacy-mode benchmark
(`synthetic-legacy-state-heavy`) added to exercise the
optimized path (the existing runes-heavy benchmark exercises a
different AST pipeline).

LOC delta: ~910 LOC net removed (1340 deleted, 430 added in
the combined module + bench).

Test plan:
- `cargo test --release --test runtime test_runtime_legacy` — 100%
- `cargo test --release --test runtime test_runtime_runes` — 100%
- `cargo test --release --test ssr` — 100%
- `cargo test --release --test compatibility_report` — 3087/3087 in-scope, every category at 100%
- `cargo fmt -- --check`, `cargo clippy --all-targets --all-features -- -D warnings`
